refactor(crypto): evict test backend from production code
- Move ALL test backend code to core/tests/backends/test_backend.c
- Production build has ZERO test symbols
- Test backend only available when GITMIND_ENABLE_TEST_BACKEND defined
- Make libsodium_backend const (immutable configuration)
- Update API to return const gm_crypto_backend_t* pointers
- Add result.h include to fix missing type warnings
- Update baseline: 353 → 345 warnings (8 eliminated)
