+++ tools/baseline_count.txt
@@ -1,1 +1,1 @@
-353
+345
